LayoutTests/imported/w3c/web-platform-tests/trusted-types/WorkerGlobalScope-eval-expected.txt


FAIL eval(string) in dedicated worker assert_throws_js: function "_ => eval("2")" did not throw
FAIL eval(TrustedScript) in dedicated worker assert_equals: expected (number) 7 but got (object) object "7"
FAIL eval(string) with default policy in dedicated worker assert_equals: expected 5 but got 2
FAIL eval(string) in shared worker assert_throws_js: function "_ => eval("2")" did not throw
FAIL eval(TrustedScript) in shared worker assert_equals: expected (number) 7 but got (object) object "7"
FAIL eval(string) with default policy in shared worker assert_equals: expected 5 but got 2
